move redux/auth/types in model User
