+++ packages/pn-personafisica-webapp/src/models/User.ts
@@ -1,0 +1,31 @@
+import { AppRouteParams, BasicUser } from '@pagopa-pn/pn-commons';
+
+export interface User extends BasicUser {
+  from_aa: boolean;
+  aud: string;
+  level: string;
+  iat: number;
+  exp: number;
+  iss: string;
+  jti: string;
+  source?: UserSource;
+}
+
+export interface UserSource {
+  channel: string; // TPP o ??? 
+  details: string;
+  retrievalId: string;
+}
+
+export interface TokenExchangeBody {
+  authorizationToken: string;
+  source?: {
+    type: 'TPP' | 'QR';
+    id: string;
+  };
+}
+
+export interface TokenExchangeRequest {
+  spidToken: string;
+  rapidAccess?: [AppRouteParams, string];
+}
