+++ packages/pn-personagiuridica-webapp/src/App.tsx
@@ -32,7 +32,7 @@ import Router from './navigation/routes';
 import * as routes from './navigation/routes.const';
 import { getCurrentAppStatus } from './redux/appStatus/actions';
 import { logout } from './redux/auth/actions';
-import { PNRole } from './redux/auth/types';
+import { PNRole } from './models/User';
 import { getDigitalAddresses } from './redux/contact/actions';
 import { useAppDispatch, useAppSelector } from './redux/hooks';
 import { getSidemenuInformation } from './redux/sidemenu/actions';
