+++ packages/pn-personagiuridica-webapp/src/components/IntegrazioneApi/VirtualKeys.tsx
@@ -24,7 +24,7 @@ import {
   deleteVirtualApiKey,
   getVirtualApiKeys,
 } from '../../redux/apikeys/actions';
-import { PNRole } from '../../redux/auth/types';
+import { PNRole } from '../../models/User';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
 import { RootState } from '../../redux/store';
 import ApiKeyModal from './ApiKeyModal';
