+++ packages/pn-personagiuridica-webapp/src/components/IntegrazioneApi/VirtualKeysTable.tsx
@@ -18,7 +18,7 @@ import {
 
 import { BffVirtualKeysResponse, VirtualKey } from '../../generated-client/pg-apikeys';
 import { ApiKeyColumnData, ModalApiKeyView } from '../../models/ApiKeys';
-import { PNRole } from '../../redux/auth/types';
+import { PNRole } from '../../models/User';
 import { useAppSelector } from '../../redux/hooks';
 import { RootState } from '../../redux/store';
 import ApiKeysDataSwitch from './ApiKeysDataSwitch';
